Фаза 0: Подготовка инфраструктуры

+++ Modules/Team/Routes/api_v1.php
@@ -10,13 +10,17 @@ use Modules\Team\Http\Controllers\V1\TeamController;
 */
 
 Route::middleware('auth:sanctum')->group(function () {
-    // Вложенные в клубы
-    Route::get('clubs/{clubId}/teams',    [TeamController::class, 'index']);
-    Route::post('clubs/{clubId}/teams',   [TeamController::class, 'store']);
 
-    // Самостоятельные
-    Route::get('teams/{id}',              [TeamController::class, 'show']);
-    Route::put('teams/{id}',              [TeamController::class, 'update']);
-    Route::delete('teams/{id}',           [TeamController::class, 'destroy']);
-    Route::post('teams/{teamId}/members', [TeamController::class, 'addMember']);
+    // Просмотр
+    Route::get('clubs/{clubId}/teams', [TeamController::class, 'index'])->middleware('permission:teams.view');
+    Route::get('teams/{id}',          [TeamController::class, 'show'])->middleware('permission:teams.view');
+
+    // Создание / редактирование / удаление
+    Route::post('clubs/{clubId}/teams', [TeamController::class, 'store'])->middleware('permission:teams.create');
+    Route::put('teams/{id}',            [TeamController::class, 'update'])->middleware('permission:teams.update');
+    Route::delete('teams/{id}',         [TeamController::class, 'destroy'])->middleware('permission:teams.delete');
+
+    // Управление составом
+    Route::post('teams/{teamId}/members', [TeamController::class, 'addMember'])
+        ->middleware('permission:teams.members');
 });
